requirements.txt
streamlit==0.79.0
pandas==1.1.3
seaborn==0.11.0
matplotlib==3.3.2
opencv_python==4.4.0.44
numpy==1.19.2
ipython==7.21.0
Pillow==8.1.2
